sanity checks, linking up files
