+++ index.html
@@ -4,10 +4,10 @@
   <meta charset="utf-8">
   <title>Weather🔊Beat</title>
 </head>
-<body id="application">
-<h4>Welcome! Please enter your location:</h4>
-  <input type="text" placeholder="Your Location">
-  <div class="application"></div>
+<body>
+<section id="application">
+
+</section>
   <script src="main.bundle.js"></script>
 </body>
 </html>
