+++ lib/app/component/LocationDisplay.jsx
