+++ lib/app/component/Summary.jsx
@@ -1,0 +1,26 @@
+const React = require('react');
+
+
+class Summary extends React.Component {
+  // constructor(props) {
+  //   super(props);
+  //   this.state = {
+  //     city: 'Default',
+  //   };
+  // }
+
+// addToCount (num) {
+//   this.setState({count: this.state.count + num});
+// }
+
+  render() {
+    return (
+  <section className="Summary">
+    <h2>A Fuckin Summary</h2>
+ </section>
+    );
+  }
+}
+
+
+module.exports = Summary;
