+++ scss/elements.scss
@@ -1,0 +1,9 @@
+@import "./variables.scss";
+
+body {
+  background: $background-color;
+}
+
+h2 {
+  font-size: 30px;
+}
