+++ scss/main.scss
@@ -1,2 +1,3 @@
 @import "./reset.scss";
-@import "./variables.scss";
+// @import "./variables.scss";
+@import "./elements.scss";
