+++ scss/variables.scss
@@ -1,4 +1,5 @@
 // colors
+$background-color:#eee43e;
 $snow-color: #eee9e9;
 $cold-color: #9ee4d9;
 $rain-color: #1733a8;
